+++ src/consts.js
@@ -1,0 +1,8 @@
+export const FILMS_COUNT = 5;
+
+export const RenderPosition = {
+  BEFORE_BEGIN: 'beforebegin',
+  AFTER_BEGIN: 'afterbegin',
+  BEFORE_END: 'beforeend',
+  AFTER_END: 'afterend',
+};
