+++ src/helpers.js
@@ -1,0 +1,5 @@
+import { RenderPosition } from './consts.js'
+
+export const renderTemplate = (container, template, place = RenderPosition.BEFORE_END) => {
+  container.insertAdjacentHTML(place, template);
+};
